+++ ChangeLog.md
@@ -117,6 +117,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 ### Added
 
 - Add `CPUX_ARGS` environment variable to set default command line
+- Add alternative key mapping for NCurses mode (option `--keymap`)
 
 ### Fixed